+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.4.1 Under development
 
+- Enh #14: Add `HasHref` trait and `href()` method to manage `href` attribute for HTML/SVG elements (@terabytesoftw)
+
 ## 0.4.0 December 27, 2025
 
 - Dep #13: Update `ui-awesome/html-helper` version constraint to `^0.6` in `composer.json` (@terabytesoftw)